+++ src/PolyJuMP.jl
@@ -79,8 +79,19 @@ function getpolyconstr(m::JuMP.Model)
     getpolydata(m).polyconstr
 end
 
+setpolymodule!(m::JuMP.Model, pm::Module) = setpolymodule!(getpolydata(m), pm)
+setpolymodule!(data::PolyData, pm::Module) = data.polymodule = pm
+
+getpolymodule(m::JuMP.Model) = getpolymodule(getpolydata(m))
+function getpolymodule(data::PolyData)
+    if isnull(data.polymodule)
+        return PolyJuMP
+    end
+    get(data.polymodule)
+end
+
 include("macros.jl")
 include("solve.jl")
-include("default_module.jl")
+include("default_methods.jl")
 
 end # module
