+++ test/polymodule.jl
@@ -2,7 +2,7 @@
     m = Model()
     # Triggers the creation of polydata
     @test isnull(PolyJuMP.getpolydata(m).polymodule)
-    @test PolyJuMP.getpolymodule(m) == PolyJuMP.DefaultModule
+    @test PolyJuMP.getpolymodule(m) == PolyJuMP
     setpolymodule!(m, TestPolyModule)
     @test PolyJuMP.getpolymodule(m) == TestPolyModule
 end
